Moved mocks to mock folder
