+++ example/AeroGearSdkExampleTests/auth/authenticator/OIDCAuthenticatorTest.swift
@@ -9,66 +9,6 @@ import XCTest
 import AppAuth
 
 class OIDCAuthenticatorTest: XCTestCase {
-    
-    class MockOIDCAuthenticator : OIDCAuthenticator {
-        
-        let fail: Bool
-        
-        init(http: AgsHttpRequestProtocol, keycloakConfig: KeycloakConfig, authConfig: AuthenticationConfig, credentialManager: CredentialManagerProtocol, fail: Bool = false) {
-            self.fail = fail
-            super.init(http: http, keycloakConfig: keycloakConfig, authConfig: authConfig, credentialManager: credentialManager)
-        }
-        
-        override func startAuthorizationFlow(byPresenting: OIDAuthorizationRequest, presenting: UIViewController, callback: @escaping (OIDCCredentials?, Error?) -> Void) -> OIDAuthorizationFlowSession {
-            var testCredential: OIDCCredentials?
-            var err: NSError?
-            
-            if (fail) {
-                testCredential = nil
-                err = NSError(domain: "errordomain", code: 123, userInfo: nil)
-            } else {
-                testCredential = OIDCCredentialsTest.buildCredentialsWithParameters(parameters: OIDCCredentialsTest.defaultParameters)
-                err = nil
-            }
-            
-            callback(testCredential, err)
-
-            return MockOIDAuthorizationFlowSession()
-        }
-    }
-    
-    class MockCredentialManager: CredentialManagerProtocol {
-        var loadCalled = false
-        var saveCalled = false
-        var clearCalled = false
-
-        func load() -> OIDCCredentials? {
-            loadCalled = true
-            return nil
-        }
-
-        func save(credentials _: OIDCCredentials) {
-            saveCalled = true
-        }
-
-        func clear() {
-            clearCalled = true
-        }
-    }
-    
-    class MockOIDAuthorizationFlowSession : NSObject, OIDAuthorizationFlowSession {
-        func cancel() {
-        }
-        func resumeAuthorizationFlow(with URL: URL) -> Bool {
-            return false
-        }
-        func failAuthorizationFlowWithError(_ error: Error) {
-        }
-    }
-    
-    class MockUIViewController : UIViewController {
-    }
-    
     var http = MockHttpRequest()
     var keycloakConfig: KeycloakConfig?
     var authConfig: AuthenticationConfig?
@@ -118,7 +58,7 @@ class OIDCAuthenticatorTest: XCTestCase {
         var onCompletedCalled = false
         let authenticator = MockOIDCAuthenticator(http: http, keycloakConfig: keycloakConfig!, authConfig: authConfig!, credentialManager: credentialManager!, fail: true)
         
-        authenticator.authenticate(presentingViewController: MockUIViewController()) {
+        authenticator.authenticate(presentingViewController: UIViewController()) {
             user, error in
             
             XCTAssertNil(user)
@@ -132,7 +72,7 @@ class OIDCAuthenticatorTest: XCTestCase {
         var onCompletedCalled = false
         let authenticator = MockOIDCAuthenticator(http: http, keycloakConfig: keycloakConfig!, authConfig: authConfig!, credentialManager: credentialManager!)
 
-        authenticator.authenticate(presentingViewController: MockUIViewController()) {
+        authenticator.authenticate(presentingViewController: UIViewController()) {
             user, error in
 
             XCTAssertNil(error)
